[inductor][reland] split test_cpp_wrapper.py into cpu and cuda test files
Summary: This relands https://github.com/pytorch/pytorch/pull/115479
after solving a test file conflict.

[ghstack-poisoned]
